I completely coded the calculator functionalitye
All buttons work and clear and process the correct math. I use switch statements to pretty up the code but im pretty sure it isnt as effecient as possible. I use 2 bool values to figure out if the 0 needs to be cleared or not or to add the numbers to the value u are typing. I also do a lot of function passing to process the onclick events on each button while still having it in the main file.